pyproject.toml
[tool.poetry]
name = "lms-system-project"
version = "0.1.0"
description = ""
authors = ["Dmitry Minin <d.n.minin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.3"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
dotenv = "^0.9.9"
pillow = "^11.2.1"
ipython = "^9.3.0"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
